requirements.txt
discord==1.7.3
discord.py==1.7.3
python-dotenv==0.19.2
pymongo==4.0.2
python-dateutil==2.7.3